fix(subprocess): if process cannot start, don't wait for it when stopping
If a child process emits `error`, the process died.  However, the `isRunning` function only checks if `proc` is `null`, so any logic in e.g., `stop()` depending on it would find a not-running process (without a `pid`!) which would apparently be running as per the `isRunning` function, and it would attempt to kill the child process and then wait for it emit `close()`.

That's not right, so now `proc` is set to `null` when it exits with an error.

Also fixed some problems in tests.