fix: handle ENOENT properly
The `errno` property of a `NodeJS.ErrnoException` is always a `number`, and `code` is what we're looking for.  At some point the behavior changed, and this looks to have been a holdover from old Node.js versions.

The relevant test made too general of an assertion, which is partly why it was not noticed.